fix: resolve all 8 CI failures in CI (core) for BpmDetector
clang-format (job 1):
  - Re-format BpmDetector.cpp, BpmDetector.hpp, test_bpm_detector.cpp
    with clang-format-18 to satisfy format checks

.jucer coherence (job 2):
  - Add dsp/src/BpmDetector.cpp to VoiceLivePro.jucer MAINGROUP

double-promotion -Werror (jobs 3-5, 7-8):
  - testing.hpp CHECK_NEAR: cast tol to double explicitly to avoid
    float→double implicit promotion at comparison site
  - test_bpm_detector.cpp: replace kBpmTol (float) literals in
    CHECK_NEAR calls — macro now handles promotion correctly

clang-tidy violations (job 6):
  - BpmDetector.cpp: add direct #include for <cstddef> <optional>
    <span> <vector> (misc-include-cleaner)
  - BpmDetector::acf() made static (readability-convert-member-to-static)
  - for (const float s : audio) const-correct loop variable
  - Fix misplaced widening casts: static_cast<size_t>(i)+static_cast<size_t>(lag)
    instead of casting the sum (bugprone-misplaced-widening-cast)
  - test: replace float loop counter pos with size_t beat index
    (cert-flp30-c / clang-analyzer-security.FloatLoopCounter)
  - test: 48000*5/48000*8 → std::size_t{48000}*N
    (bugprone-implicit-widening-of-multiplication-result)
  - test: hoist OR into bool nearExpected to avoid DeMorgan warning
    inside CHECK macro expansion (readability-simplify-boolean-expr)

+++ testing/include/voicelive_testing/testing.hpp
@@ -101,7 +101,7 @@ inline int runAll() {
     do {                                                                                        \
         const double vl_a = static_cast<double>(actual);                                        \
         const double vl_e = static_cast<double>(expected);                                      \
-        if (!(std::abs(vl_a - vl_e) <= (tol))) {                                                \
+        if (!(std::abs(vl_a - vl_e) <= static_cast<double>(tol))) {                                                \
             ++::vltest::currentFailures();                                                      \
             std::cout << "  CHECK_NEAR échec : " #actual " ≈ " #expected " (" << vl_a << " vs " \
                       << vl_e << ", tol " << (tol) << ")  (" << __FILE__ << ":" << __LINE__     \
